DE/requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
dotenv==0.9.9
finance-datareader==0.9.96
greenlet==3.1.1
idna==3.10
lxml==5.3.2
narwhals==1.34.0
numpy==2.0.2
packaging==24.2
pandas==2.2.3
plotly==6.0.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-file==2.1.0
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.40
tqdm==4.67.1
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
